+++ R/ml_summary.R
@@ -9,7 +9,21 @@ new_ml_summary <- function(jobj, ..., class = character()) {
   )
 }
 
-new_ml_summary_clustering <- function(jobj, ..., class = character()) {
+#' @export
+print.ml_summary <- function(x, ...) {
+  short_type <- strsplit(x$type, "\\.") %>%
+    unlist() %>%
+    dplyr::last()
+
+  cat(short_type, "\n")
+  cat(" Access the following via `$` or `ml_summary()`.", "\n")
+  for (m in setdiff(names(x), c("type", ".jobj"))) {
+    print_value <- if (is.function(x[[m]])) paste0(m, "()") else m
+    cat(" -", print_value, "\n")
+  }
+}
+
+new_ml_clustering_summary <- function(jobj, ..., class = character()) {
   new_ml_summary(
     jobj,
     cluster = function() invoke(jobj, "cluster") %>% sdf_register(), # lazy val
@@ -18,6 +32,7 @@ new_ml_summary_clustering <- function(jobj, ..., class = character()) {
     k = invoke(jobj, "k"),
     prediction_col = invoke(jobj, "predictionCol"),
     predictions = invoke(jobj, "predictions") %>% sdf_register(),
-    class = c(class, "ml_summary_clustering")
+    ...,
+    class = c(class, "ml_clustering_summary")
   )
 }