+++ R/shell_connection.R
@@ -241,16 +241,13 @@ start_shell <- function(master,
              gsub("[-_a-zA-Z]", "", spark_version)
       )
     )
-    if (spark_version < "2.0")
-      scala_version <- numeric_version("2.10")
-    else
-      scala_version <- numeric_version("2.11")
-    extensions <- spark_dependencies_from_extensions(spark_version, scala_version, extensions)
+    extensions <- spark_dependencies_from_extensions(spark_version, extensions, config)
 
     # combine passed jars and packages with extensions
     all_jars <- c(jars, extensions$jars)
     jars <- if (length(all_jars) > 0) normalizePath(unlist(unique(all_jars))) else list()
     packages <- unique(c(packages, extensions$packages))
+    repositories <- extensions$repositories
 
     # include embedded jars, if needed
     csv_config_value <- spark_config_value(config, c("sparklyr.connect.csv.embedded", "sparklyr.csv.embedded"))
@@ -282,6 +279,11 @@ start_shell <- function(master,
     # add packages to arguments
     if (length(packages) > 0) {
       shell_args <- c(shell_args, "--packages", paste(shQuote(packages, type = shQuoteType), collapse=","))
+    }
+
+    # add repositories to arguments
+    if (length(repositories) > 0) {
+      shell_args <- c(shell_args, "--repositories", paste(repositories, collapse=","))
     }
 
     # add environment parameters to arguments
@@ -428,6 +430,7 @@ start_shell <- function(master,
     app_name = app_name,
     config = config,
     state = new.env(),
+    extensions = extensions,
     # spark_shell_connection
     spark_home = spark_home,
     backend = backend,