+++ R/spark_extensions.R
@@ -28,35 +28,71 @@ registered_extensions <- function() {
 #'
 #' Define a Spark dependency consisting of a set of custom JARs and Spark packages.
 #'
-#' @param jars Character vector of full paths to JAR files
-#' @param packages Character vector of Spark packages names
+#' @param jars Character vector of full paths to JAR files.
+#' @param packages Character vector of Spark packages names.
+#' @param initializer Optional callback function called when initializing a connection.
+#' @param catalog Optional location where extension JAR files can be downloaded for Livy.
+#' @param repositories Character vector of Spark package repositories.
+#' @param ... Additional optional arguments.
 #'
 #' @return An object of type `spark_dependency`
 #'
 #' @export
-spark_dependency <- function(jars = NULL, packages = NULL) {
+spark_dependency <- function(jars = NULL,
+                             packages = NULL,
+                             initializer = NULL,
+                             catalog = NULL,
+                             repositories = NULL,
+                             ...) {
   structure(class = "spark_dependency", list(
     jars = jars,
-    packages = packages
+    packages = packages,
+    initializer = initializer,
+    catalog = catalog,
+    repositories = repositories
   ))
 }
 
-spark_dependencies_from_extensions <- function(spark_version, scala_version, extensions) {
+spark_dependencies_from_extensions <- function(spark_version, extensions, config) {
+  if (spark_version < "2.0")
+    scala_version <- numeric_version("2.10")
+  else
+    scala_version <- numeric_version("2.11")
 
   jars <- character()
   packages <- character()
+  initializers <- list()
+  catalog_jars <- character()
+  repositories <- character()
 
-  lapply(extensions, function(extension) {
+  for (extension in extensions) {
     dependencies <- spark_dependencies_from_extension(spark_version, scala_version, extension)
-    lapply(dependencies, function(dependency) {
-      jars <<- c(jars, dependency$jars)
-      packages <<- c(packages, dependency$packages)
-    })
-  })
+    for (dependency in dependencies) {
+      jars <- c(jars, dependency$jars)
+      packages <- c(packages, dependency$packages)
+      initializers <- c(initializers, dependency$initializer)
+      repositories <- c(repositories, dependency$repositories)
+
+      config_catalog <- spark_config_value(config, "sparklyr.extensions.catalog", TRUE)
+      if (!identical(dependency$catalog, NULL) && !identical(config_catalog, FALSE)) {
+        catalog_path <- dependency$catalog
+        if (is.character(config_catalog)) catalog_path <- config_catalog
+
+        if (!grepl("%s", config_catalog)) config_catalog <- paste0(config_catalog, "%s")
+
+        catalog_jars <- c(
+          catalog_jars,
+          sprintf(config_catalog, basename(jars))
+        )
+      }
+    }
+  }
 
   list(
     jars = jars,
-    packages = packages
+    packages = packages,
+    initializers = initializers,
+    catalog_jars = catalog_jars
   )
 }
 
@@ -87,6 +123,21 @@ spark_dependencies_from_extension <- function(spark_version, scala_version, exte
 
   # return it
   dependency
+}
+
+#' Fallback to Spark Dependency
+#'
+#' Helper function to assist falling back to previous Spark versions.
+#'
+#' @param spark_version The Spark version being requested in \code{spark_dependencies}.
+#' @param supported_versions The Spark versions that are supported by this extension.
+#'
+#' @return A Spark version to use.
+#'
+#' @export
+spark_dependency_fallback <- function(spark_version, supported_versions) {
+  supported <- supported_versions[package_version(supported_versions) <= spark_version]
+  sort(supported, decreasing = TRUE)[[1]]
 }
 
 sparklyr_jar_path <- function(spark_version) {